+++ src/zm_font.cpp
@@ -101,7 +101,7 @@ FontLoadError ZmFont::LoadFontFile(const std::string &loc) {
     }
 
     std::vector<uint64> bitmap;
-    bitmap.resize(bitmap_header.number_of_code_points * bitmap_header.char_height);
+    bitmap.resize(static_cast<std::size_t>(bitmap_header.number_of_code_points) * bitmap_header.char_height);
 
     std::size_t bitmap_bytes = bitmap.size() * sizeof(uint64);
     font_file.read(reinterpret_cast<char *>(bitmap.data()), static_cast<std::streamsize>(bitmap_bytes));
